Create converter_data.h
Header file for helper functions in converter project. I put the definitions of the helper functions in here because putting them in a separate *.cpp file gave me linker errors I couldn't figure out (specifically, it said something about defining a symbol twice...). I put everything inside a namespace to try and fix this, but it didn't work, so in the end I just dumped it all here.

This also contains the addition of the alternateNamesForMassUnits map, which allows for multiple ways of specifying the same thing.